refactor(cli): migrate eumetsat tooling onto discovery
Move eumetsat's refresher/writer/curated_ids/prober/emitter to
earthlens.eumetsat.cli and drop them (with the browse-URL constant) from core's
refresh/curate/stanza. Relocate its CLI tests to earthlens-imagery. Shrink the
guard allow-list (eumetsat).
